+++ data/vocab_usage.json
@@ -135,7 +135,82 @@
     "peek": 1,
     "open": 1,
     "relax": 1,
-    "set": 1
+    "set": 1,
+    "receive": 0,
+    "dare": 0,
+    "follow": 0,
+    "should": 0,
+    "mind": 0,
+    "excuse": 0,
+    "remember": 0,
+    "release": 0,
+    "return": 0,
+    "guess": 0,
+    "step": 0,
+    "knit": 0,
+    "starve": 0,
+    "count": 0,
+    "laugh": 0,
+    "wish": 0,
+    "would": 0,
+    "tighten": 0,
+    "apply": 0,
+    "touch": 0,
+    "change": 0,
+    "swear": 0,
+    "win": 0,
+    "jump": 0,
+    "happen": 0,
+    "allow": 0,
+    "handle": 0,
+    "may": 0,
+    "remove": 0,
+    "kiss": 0,
+    "hang": 0,
+    "forgive": 0,
+    "draw": 0,
+    "suggest": 0,
+    "insist": 0,
+    "care": 0,
+    "recognize": 0,
+    "perceive": 0,
+    "pull": 0,
+    "search": 0,
+    "calm": 0,
+    "stick": 0,
+    "shit": 0,
+    "manage": 0,
+    "getting": 0,
+    "cover": 0,
+    "join": 0,
+    "bring": 0,
+    "drive": 0,
+    "fix": 0,
+    "resign": 0,
+    "require": 0,
+    "'": 0,
+    "pass": 0,
+    "sound": 0,
+    "please": 0,
+    "marry": 0,
+    "appreciate": 0,
+    "depend": 0,
+    "prove": 0,
+    "answer": 0,
+    "apologize": 0,
+    "show": 0,
+    "bet": 0,
+    "welcome": 0,
+    "believe": 0,
+    "end": 0,
+    "trust": 0,
+    "could": 0,
+    "hold": 0,
+    "continue": 0,
+    "joke": 0,
+    "promise": 0,
+    "enter": 0,
+    "explain": 0
   },
   "vocab": {
     "about": 23,
@@ -549,6 +624,173 @@
     "place": 1,
     "glad": 1,
     "anymore": 1,
-    "those": 1
+    "those": 1,
+    "late": 0,
+    "honor": 0,
+    "correct": 0,
+    "preferably": 0,
+    "actually": 0,
+    "true": 0,
+    "favor": 0,
+    "fast": 0,
+    "amazing": 0,
+    "darling": 0,
+    "goodness": 0,
+    "forever": 0,
+    "quite": 0,
+    "indeed": 0,
+    "christmas": 0,
+    "gentleman": 0,
+    "honey": 0,
+    "doctor": 0,
+    "afternoon": 0,
+    "pleasure": 0,
+    "pregnant": 0,
+    "fun": 0,
+    "deep": 0,
+    "nicely": 0,
+    "point": 0,
+    "twice": 0,
+    "huh": 0,
+    "little": 0,
+    "against": 0,
+    "breath": 0,
+    "pen": 0,
+    "gear": 0,
+    "mr.": 0,
+    "ma'am": 0,
+    "necessarily": 0,
+    "reason": 0,
+    "nine": 0,
+    "dressed": 0,
+    "quit": 0,
+    "highness": 0,
+    "unfair": 0,
+    "name": 0,
+    "captain": 0,
+    "awake": 0,
+    "gladly": 0,
+    "dad": 0,
+    "married": 0,
+    "agree": 0,
+    "prepared": 0,
+    "calm": 0,
+    "sweetie": 0,
+    "hand": 0,
+    "mine": 0,
+    "order": 0,
+    "nowhere": 0,
+    "ill": 0,
+    "hug": 0,
+    "t": 0,
+    "son": 0,
+    "later": 0,
+    "cute": 0,
+    "forward": 0,
+    "around": 0,
+    "genius": 0,
+    "anyone": 0,
+    "else": 0,
+    "directly": 0,
+    "almost": 0,
+    "carefully": 0,
+    "nice": 0,
+    "birthday": 0,
+    "far": 0,
+    "boy": 0,
+    "any": 0,
+    "clear": 0,
+    "eye": 0,
+    "message": 0,
+    "merry": 0,
+    "hey": 0,
+    "tasty": 0,
+    "minute": 0,
+    "gift": 0,
+    "chance": 0,
+    "unharmed": 0,
+    "none": 0,
+    "lady": 0,
+    "ten": 0,
+    "care": 0,
+    "accident": 0,
+    "hi": 0,
+    "small": 0,
+    "woman": 0,
+    "mouth": 0,
+    "miracle": 0,
+    "guy": 0,
+    "daughter": 0,
+    "ride": 0,
+    "enough": 0,
+    "mean": 0,
+    "husband": 0,
+    "relax": 0,
+    "drink": 0,
+    "majesty": 0,
+    "simple": 0,
+    "end": 0,
+    "plate": 0,
+    "open": 0,
+    "handsome": 0,
+    "everything": 0,
+    "honest": 0,
+    "sir": 0,
+    "dear": 0,
+    "quiet": 0,
+    "mom": 0,
+    "wonderful": 0,
+    "fan": 0,
+    "suggestion": 0,
+    "boss": 0,
+    "these": 0,
+    "follow": 0,
+    "thousand": 0,
+    "anyway": 0,
+    "usual": 0,
+    "cop": 0,
+    "yours": 0,
+    "buddy": 0,
+    "min": 0,
+    "start": 0,
+    "whom": 0,
+    "seriously": 0,
+    "oops": 0,
+    "god": 0,
+    "incredible": 0,
+    "ago": 0,
+    "busy": 0,
+    "whatever": 0,
+    "luck": 0,
+    "once": 0,
+    "stay": 0,
+    "everyone": 0,
+    "even": 0,
+    "neither": 0,
+    "along": 0,
+    "special": 0,
+    "possible": 0,
+    "aboard": 0,
+    "enter": 0,
+    "ahead": 0,
+    "patient": 0,
+    "certainly": 0,
+    "concentrate": 0,
+    "same": 0,
+    "pretty": 0,
+    "life": 0,
+    "police": 0,
+    "492": 0,
+    "move": 0,
+    "second": 0,
+    "turn": 0,
+    "still": 0,
+    "someone": 0,
+    "news": 0,
+    "moment": 0,
+    "plan": 0,
+    "choice": 0,
+    "impressed": 0,
+    "try": 0
   }
 }
